untangle `fix-regexp-well-known-symbol-logic` by moving feature detection to related modules
